Prerelease

+++ Controllers/IngredienteController.cs
@@ -12,69 +12,97 @@ namespace CupMetric.Controllers
         {
             this.repo = repo;
         }
-        // GET: IngredienteController
         public async Task<ActionResult> List()
         {
+            string rol = HttpContext.Session.GetString("IDROL");
+            if (rol == null)
+            {
+                return RedirectToAction("Index", "Home");
+            }
+            int idRol = int.Parse(HttpContext.Session.GetString("IDROL"));
+            if (idRol != 2)
+            {
+                return RedirectToAction("Index", "Home");
+            }
             List<Ingrediente> ingredientes = await this.repo.GetIngredientesAsync();
             return View(ingredientes);
         }
 
-        // GET: IngredienteController/Details/5
         public async Task<ActionResult> Details(int idIngrediente)
         {
+            string rol = HttpContext.Session.GetString("IDROL");
+            if (rol == null)
+            {
+                return RedirectToAction("Index", "Home");
+            }
+            int idRol = int.Parse(HttpContext.Session.GetString("IDROL"));
+            if (idRol != 2)
+            {
+                return RedirectToAction("Index", "Home");
+            }
             Ingrediente ingrediente = await this.repo.FindIngredienteByIdAsync(idIngrediente);
             return View(ingrediente);
         }
 
-        // GET: IngredienteController/Create
         public ActionResult Create()
         {
+            string rol = HttpContext.Session.GetString("IDROL");
+            if (rol == null)
+            {
+                return RedirectToAction("Index", "Home");
+            }
+            int idRol = int.Parse(HttpContext.Session.GetString("IDROL"));
+            if (idRol != 2)
+            {
+                return RedirectToAction("Index", "Home");
+            }
             return View();
         }
 
-        // POST: IngredienteController/Create
         [HttpPost]
         [ValidateAntiForgeryToken]
         public async Task<ActionResult> Create(Ingrediente ingrediente)
         {
-            try
-            {
                 await this.repo.CreateIngredienteAsync(ingrediente);
                 return RedirectToAction("List");
-            }
-            catch
-            {
-                return View();
-            }
         }
 
         public async Task<ActionResult> Update(int idingrediente)
         {
+            string rol = HttpContext.Session.GetString("IDROL");
+            if (rol == null)
+            {
+                return RedirectToAction("Index", "Home");
+            }
+            int idRol = int.Parse(HttpContext.Session.GetString("IDROL"));
+            if (idRol != 2)
+            {
+                return RedirectToAction("Index", "Home");
+            }
             Ingrediente ingrediente = await this.repo.FindIngredienteByIdAsync(idingrediente);
             return View(ingrediente);
         }
 
-        // POST: IngredienteController/Edit/5
         [HttpPost]
-        [ValidateAntiForgeryToken]
         public async Task<ActionResult> Update(Ingrediente ingrediente)
         {
-            try
-            {
-                await this.repo.UpdateIngredienteAsync(ingrediente);
+            await this.repo.UpdateIngredienteAsync(ingrediente);
                 return RedirectToAction("List");
-            }
-            catch
-            {
-                return View();
-            }
         }
-
-        // GET: IngredienteController/Delete/5
         public async Task<ActionResult> Delete(int idingrediente)
         {
-            this.repo.DeleteIngredienteAsync(idingrediente);
-            return View();
+            string rol = HttpContext.Session.GetString("IDROL");
+            if (rol == null)
+            {
+                return RedirectToAction("Index", "Home");
+            }
+            int idRol = int.Parse(HttpContext.Session.GetString("IDROL"));
+            if (idRol != 2)
+            {
+                return RedirectToAction("Index", "Home");
+            }
+            await this.repo.DeleteIngredienteAsync(idingrediente);
+            return RedirectToAction("List");
         }
     }
 }
